lint scripts and refactor profiling
