+++ .pre-commit-config.yaml
@@ -1,4 +1,4 @@
-exclude: scripts/
+#exclude: scripts/
 repos:
 -   repo: https://github.com/pre-commit/pre-commit-hooks
     rev: v2.3.0
